extract ofxImGuiLoggerChannel
